Allow removing of customers with threads
As soon as a customer started a thread it is impossible to remove them
and also the corresponding employee. Also all employees who have
answered in that thread cannot be removed, as the messages cannot be
removed.

Now the database is cascading the removal of a customer down and it
removes a thread and the messages. It notifies the admin that it will do
so.